Create Linked-Lists: Detect-A-Cycle
Detects if there is a cycle in a Linked List given a list less than 100. If the list was larger than 100 or we were unaware of its size, I would suggest using two nodes (tortoise and hare) where the tortoise moves by 1 node per step, while the hare moves at 2. Both of them would start at the head. If they were ever equal after that, we would know there's a cycle since the hare looped back and caught up to the tortoise again.